refactor: CreateEvent component
