--- src/components/DynamicField/index.js
@@ -1,3 +1,0 @@
-import DynamicField from './DynamicField';
-
-export default DynamicField;
