+++ src/components/EventEditor/EditEvent/index.js
