+++ src/components/EventEditor/FormComponents/CommentField/index.js
@@ -1,0 +1,3 @@
+import CommentField from './CommentField';
+
+export default CommentField;
