+++ src/components/EventEditor/FormComponents/DateTimeComponent/index.js
@@ -1,0 +1,3 @@
+import DateTimeComponent from './DateTimeComponent';
+
+export default DateTimeComponent;
