+++ src/components/EventEditor/FormComponents/DeleteOrganizerButton/index.js
