+++ src/components/EventEditor/FormComponents/DemoUrlField/index.js
@@ -1,0 +1,3 @@
+import DemoUrlField from './DemoUrlField';
+
+export default DemoUrlField;
