+++ src/components/EventEditor/FormComponents/DescriptionUrlField/index.js
@@ -1,0 +1,3 @@
+import DescriptionUrlField from './DescriptionUrlField';
+
+export default DescriptionUrlField;
