+++ src/components/EventEditor/FormComponents/MaterialsDynamicField/FieldTemplate/index.js
