+++ src/components/EventEditor/FormComponents/MaterialsDynamicField/index.js
@@ -1,0 +1,3 @@
+import MaterialsDynamicField from './MaterialsDynamicField';
+
+export default MaterialsDynamicField;
