+++ src/components/EventEditor/FormComponents/OrganizerSelect/index.js
@@ -1,0 +1,3 @@
+import OrganizerSelect from './OrganizerSelect';
+
+export default OrganizerSelect;
