+++ src/components/EventEditor/FormComponents/PlaceComponent/index.js
@@ -1,0 +1,3 @@
+import PlaceComponent from './PlaceComponent';
+
+export default PlaceComponent;
