+++ src/components/EventEditor/FormComponents/ScreenUrlField/index.js
@@ -1,0 +1,3 @@
+import ScreenUrlField from './ScreenUrlField';
+
+export default ScreenUrlField;
