+++ src/components/EventEditor/FormComponents/TimeZoneSelect/index.js
@@ -1,0 +1,3 @@
+import TimeZoneSelect from './TimeZoneSelect';
+
+export default TimeZoneSelect;
