+++ src/components/EventEditor/FormComponents/TopicField/index.js
@@ -1,0 +1,3 @@
+import TopicField from './TopicField';
+
+export default TopicField;
