+++ src/components/EventEditor/FormComponents/TypeSelect/index.js
@@ -1,0 +1,3 @@
+import TypeSelect from './TypeSelect';
+
+export default TypeSelect;
