+++ src/components/EventEditor/FormComponents/index.js
@@ -1,0 +1,29 @@
+import DeleteOrganizerButton from './DeleteOrganizerButton';
+import DescriptionUrlField from './DescriptionUrlField';
+import MaterialsDynamicField from './MaterialsDynamicField';
+import OrganizerSelect from './OrganizerSelect';
+import TimeZoneSelect from './TimeZoneSelect';
+import TopicField from './TopicField';
+import TypeSelect from './TypeSelect';
+import DateTimeComponent from './DateTimeComponent';
+import PlaceComponent from './PlaceComponent';
+import DemoUrlField from './DemoUrlField';
+import ScreenUrlField from './ScreenUrlField';
+import DescriptionComponent from './DescriptionComponent';
+import CommentField from './CommentField';
+
+export {
+  DeleteOrganizerButton,
+  MaterialsDynamicField,
+  TimeZoneSelect,
+  TopicField,
+  DescriptionUrlField,
+  OrganizerSelect,
+  TypeSelect,
+  DateTimeComponent,
+  PlaceComponent,
+  DemoUrlField,
+  ScreenUrlField,
+  DescriptionComponent,
+  CommentField,
+};
