+++ src/components/EventEditor/index.js
@@ -1,0 +1,4 @@
+import CreateEvent from './CreateEvent';
+import EditEvent from './EditEvent';
+
+export {CreateEvent, EditEvent};
