+++ src/constants/event-object.js
@@ -1,0 +1,20 @@
+import {DEFAULT_TIMEZONE} from './timezones';
+
+export default {
+  id: '',
+  topic: '',
+  description: '',
+  descriptionUrl: '',
+  type: '',
+  timeZone: DEFAULT_TIMEZONE,
+  dateTime: 0,
+  taskObj: {
+    demoUrl: '',
+    materials: [],
+    screen: '',
+  },
+  place: '',
+  comment: '',
+  organizer: '',
+  feedback: false,
+};
